Remove RmqCoordinator to tests/util only
