+++ src/plumpy/message.py
@@ -10,8 +10,6 @@ from typing import TYPE_CHECKING, Any, Dict, Optional, Sequence, Union, cast
 from plumpy.coordinator import Coordinator
 from plumpy.exceptions import PersistenceError, TaskRejectedError
 
-from plumpy.exceptions import PersistenceError, TaskRejectedError
-
 from . import loaders, persistence
 from .utils import PID_TYPE
